requirements.txt
discord.py==1.7.1
beautifulsoup4
requests
PyDictionary